Abstracted out InputListener in prep for keyboard input
